+++ R/itemplot.internal.R
@@ -72,19 +72,22 @@ setMethod(
         dat2$cat <- rep(as.character(1:(length(Plist))), each = nrow(dat))
         if(all(dat2$cat == '0')) dat2$cat <- rep('1', length(dat2$cat))
         if(nfact == 1){
-            if(type == 'info')
+            if(type == 'info'){
                 return(lattice::xyplot(info ~ Theta, dat, group=group, type = 'l',
                                        auto.key = TRUE, main = paste('Information for item', item),
                                        ylab = expression(I(theta)), xlab = expression(theta), ...))
-            if(type == 'trace')
+            } else if(type == 'trace'){
                 return(lattice::xyplot(P  ~ Theta | cat, dat2, group=group, type = 'l',
                             auto.key = TRUE, main = paste("Item", item, "Trace"), ylim = c(-0.1,1.1),
                             ylab = expression(P(theta)), xlab = expression(theta), ...))
-            if(type == 'RE')
+            } else if(type == 'RE'){
                 return(lattice::xyplot(info ~ Theta, dat, group=group, type = 'l',
                                        auto.key = TRUE,
                                        main = paste('Relative efficiency for item', item),
                                        ylab = expression(RE(theta)), xlab = expression(theta), ...))
+            } else {
+                stop('Plot type not supported for unidimensional model')
+            }
         }
         if(nfact == 2){
             Names <- colnames(dat)
@@ -93,14 +96,14 @@ setMethod(
             Names2[2:3] <- c('Theta2', 'Theta1')
             colnames(dat) <- Names
             colnames(dat2) <- Names2
-            if(type == 'info')
+            if(type == 'info'){
                 return(lattice::wireframe(info ~ Theta1 + Theta2, data = dat, group=group,
                                           main=paste("Item", item, "Information"),
                                           zlab=expression(I(theta)), xlab=expression(theta[1]),
                                           ylab=expression(theta[2]), screen=rot,
                                           scales = list(arrows = FALSE),
                                           auto.key = TRUE, ...))
-            if(type == 'trace')
+            } else if(type == 'trace'){
                 return(lattice::wireframe(P ~ Theta1 + Theta2|cat, data = dat2, group = group,
                                           main = paste("Item", item, "Trace"),
                                           zlab=expression(P(theta)),
@@ -108,13 +111,16 @@ setMethod(
                                           ylab=expression(theta[2]), zlim = c(-0.1,1.1),
                                           scales = list(arrows = FALSE), screen=rot,
                                           auto.key = TRUE, ...))
-            if(type == 'RE')
+            } else if(type == 'RE'){
                 return(lattice::wireframe(info ~ Theta1 + Theta2, data = dat, group=group,
                                           main=paste("Relative efficiency for item", item),
                                           zlab=expression(RE(theta)), xlab=expression(theta[1]),
                                           ylab=expression(theta[2]),
                                           scales = list(arrows = FALSE), screen=rot,
                                           auto.key = TRUE, ...))
+            } else {
+                stop('Plot type not supported for 2 dimensional model')
+            }
         }
     }
 )
@@ -123,9 +129,10 @@ setMethod(
 itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zeros, rot, ...){
     if(drop.zeros) x@pars[[item]] <- extract.item(x, item, drop.zeros=TRUE)
     nfact <- min(x@pars[[item]]@nfact, x@nfact)
-    if(nfact > 2) stop('Can not plot high dimensional models')
+    if(nfact > 3) stop('Can not plot high dimensional models')
     if(nfact == 2 && is.null(degrees)) stop('Please specify a vector of angles that sum to 90')
     theta <- seq(-4,4, length.out=40)    
+    if(nfact == 3) theta <- seq(-4,4, length.out=20)    
     prodlist <- attr(x@pars, 'prodlist')    
     if(length(prodlist) > 0){        
         Theta <- thetaComb(theta, x@nfact)        
@@ -141,6 +148,13 @@ itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zer
     if(is(x@pars[[item]], 'custom') && any(type %in% c('info', 'infocontour')))
         stop('Unable to compute information for custom items')
     if(!class(x@pars[[item]]) %in% c('custom')){
+        if(nfact == 3){
+            if(length(degrees) != 3 && any(type %in% 'info', 'SE')){
+                warning('Information plots require the degrees input to be of length 3')
+            } else {
+                info <- iteminfo(x=x@pars[[item]], Theta=ThetaFull, degrees=degrees)
+            } 
+        }
         if(nfact == 2){
             for(i in 1:length(degrees))
                 info <- info + iteminfo(x=x@pars[[item]], Theta=ThetaFull, degrees=c(degrees[i],
@@ -151,7 +165,7 @@ itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zer
     } else message('Information functions could not be computed')    
     CEinfoupper <- CEinfolower <- info
     CEprobupper <- CEproblower <- P
-    if(CE){
+    if(CE && nfact != 3){
         tmpitem <- x@pars[[item]]
         if(length(tmpitem@SEpar) == 0) stop('Must calculate the information matrix first.')
         splt <- strsplit(colnames(x@information), '\\.')
@@ -222,8 +236,7 @@ itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zer
                 return(lattice::xyplot(P ~ Theta, plt2, group = time, type = 'l', auto.key = TRUE,
                                 main = paste('Trace lines for item', item), ylim = c(-0.1,1.1),
                                 ylab = expression(P(theta)), xlab = expression(theta), ... ))
-        }
-        if(type == 'info'){
+        } else if(type == 'info'){
             if(CE){
                 return(lattice::xyplot(info + CEinfoupper + CEinfolower ~ Theta, plt, type = 'l',
                                 col = c('black', 'red', 'red'), lty = c(1,2,2),
@@ -233,27 +246,27 @@ itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zer
                 return(lattice::xyplot(info ~ Theta, plt, type = 'l',
                                 auto.key = TRUE, main = paste('Information for item', item),
                                 ylab = expression(I(theta)), xlab = expression(theta), ...))
-        }
-        if(type == 'score'){
+        } else if(type == 'score'){
             return(lattice::xyplot(score ~ Theta, plt, type = 'l',
                             auto.key = TRUE, main = paste('Expected score for item', item),
                             ylab = expression(E(theta)), xlab = expression(theta), ...))
-        }
-        if(type == 'SE'){
+        } else if(type == 'SE'){
             return(lattice::xyplot(SE ~ Theta, plt, type = 'l',
                                    auto.key = TRUE, main = paste('Standard error plot for item', item),
                                    ylab = expression(SE(theta)), xlab = expression(theta), ...))
-        }
-        if(type == 'infoSE'){
+        } else if(type == 'infoSE'){
             obj1 <- xyplot(info~Theta, plt, type='l',
                            main = paste('Item information and standard errors for item', item),
                            xlab = expression(theta), ylab=expression(I(theta)))
             obj2 <- xyplot(SE~Theta, plt, type='l', ylab=expression(SE(theta)))
             if(!require(latticeExtra)) require(latticeExtra)
             return(doubleYScale(obj1, obj2, add.ylab2 = TRUE))
-        }
-        if(type == 'infocontour') stop('Cannot draw contours for 1 factor models')
-    } else {
+        } else if(type == 'infocontour'){
+            stop('Cannot draw contours for 1 factor models')
+        } else {
+            stop('Plot type not supported for unidimensional model')
+        }        
+    } else if(nfact == 2){
         plt <- data.frame(info = info, SE = 1/sqrt(info), Theta1 = Theta[,1], Theta2 = Theta[,2])
         plt2 <- data.frame(P = P, Theta1 = Theta[,1], Theta2 = Theta[,2])
         colnames(plt2) <- c(paste("P", 1:ncol(P), sep=''), "Theta1", "Theta2")
@@ -264,11 +277,11 @@ itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zer
         plt$CEinfolower <- CEinfolower
         plt2$upper <- as.numeric(CEprobupper)
         plt2$lower <- as.numeric(CEproblower)
-        if(type == 'infocontour')
+        if(type == 'infocontour'){
             return(contourplot(info ~ Theta1 * Theta2, data = plt,
                                main = paste("Item", item, "Information Contour"), xlab = expression(theta[1]),
                                ylab = expression(theta[2]), ...))
-        if(type == 'info')
+        } else if(type == 'info'){
             if(CE)
                 return(lattice::wireframe(info + CEinfolower + CEinfoupper ~ Theta1 + Theta2, data = plt,
                                    main = paste("Item", item, "Information"), col = c('black', 'red', 'red'),
@@ -279,7 +292,7 @@ itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zer
                              main = paste("Item", item, "Information"),
                              zlab=expression(I(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
                              scales = list(arrows = FALSE), colorkey = TRUE, drape = TRUE, screen=rot, ...))
-        if(type == 'trace'){
+        } else if(type == 'trace'){
             if(CE)
                 return(lattice::wireframe(P + upper + lower ~ Theta1 + Theta2 | time, data = plt2,
                                           main = paste("Item", item, "Trace"), zlim = c(-0.1,1.1),
@@ -292,18 +305,50 @@ itemplot.main <- function(x, item, type, degrees, CE, CEalpha, CEdraws, drop.zer
                              main = paste("Item", item, "Trace"), zlim = c(-0.1,1.1),
                              zlab=expression(P(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
                              scales = list(arrows = FALSE), colorkey = TRUE, drape = TRUE, screen=rot, ...))
-        }
-        if(type == 'score'){
+        } else if(type == 'score'){
             return(lattice::wireframe(score ~ Theta1 + Theta2, data = plt, main = paste("Item", item, "Expected Scores"),
                                       zlab=expression(E(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
                                       zlim = c(min(floor(plt$score)), max(ceiling(plt$score))),scales = list(arrows = FALSE),
                                       colorkey = TRUE, drape = TRUE, screen=rot, ...))
-        }
-        if(type == 'SE'){
+        } else if(type == 'SE'){
             return(lattice::wireframe(SE ~ Theta1 + Theta2, data = plt, main = paste("Item", item, "Standard Errors"),
                                       zlab=expression(SE(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
                                       scales = list(arrows = FALSE),
                                       colorkey = TRUE, drape = TRUE, screen=rot, ...))
+        } else {
+            stop('Plot type not supported for 2 dimensional model')
         }
+    } else {        
+        plt <- data.frame(info = info, SE = 1/sqrt(info), Theta1 = Theta[,1], Theta2 = Theta[,2], 
+                          Theta3 = Theta[,3])
+        plt2 <- data.frame(P = P, Theta1 = Theta[,1], Theta2 = Theta[,2], Theta3 = Theta[,3])
+        colnames(plt2) <- c(paste("P", 1:ncol(P), sep=''), "Theta1", "Theta2", "Theta3")
+        plt2 <- reshape(plt2, direction='long', varying = paste("P", 1:ncol(P), sep=''), v.names = 'P',
+                        times = paste("P", 1:ncol(P), sep=''))
+        plt$score <- score        
+        if(type == 'trace'){
+            return(lattice::wireframe(P ~ Theta1 + Theta2|Theta3, data = plt2, group = time,
+                                      main = paste("Item", item, "Trace"), zlim = c(-0.1,1.1),
+                                      zlab=expression(P(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
+                                      scales = list(arrows = FALSE), colorkey = TRUE, drape = TRUE, screen=rot, ...))
+        } else if(type == 'score'){
+            return(lattice::wireframe(score ~ Theta1 + Theta2|Theta3, data = plt, main = paste("Item", item, "Expected Scores"),
+                                      zlab=expression(E(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
+                                      zlim = c(min(floor(plt$score)), max(ceiling(plt$score))),scales = list(arrows = FALSE),
+                                      colorkey = TRUE, drape = TRUE, screen=rot, ...))
+        } else if(type == 'info'){
+            return(lattice::wireframe(info ~ Theta1 + Theta2|Theta3, data = plt,
+                                      main = paste("Item", item, "Information"),
+                                      zlab=expression(I(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
+                                      scales = list(arrows = FALSE), colorkey = TRUE, drape = TRUE, screen=rot, ...))            
+        } else if(type == 'SE'){
+            return(lattice::wireframe(SE ~ Theta1 + Theta2|Theta3, data = plt, main = paste("Item", item, "Standard Errors"),
+                                      zlab=expression(SE(theta)), xlab=expression(theta[1]), ylab=expression(theta[2]),
+                                      scales = list(arrows = FALSE),
+                                      colorkey = TRUE, drape = TRUE, screen=rot, ...))
+        } else {
+            stop('Plot type not supported for 3 dimensional model')            
+        }
+        
     }
 }